Feature/wordpress to md (#27)
Add the ability to convert a WordPress export file into n markdown files representing each blog post (published and draft). Metadata (categories, tags, author info) is contained in YAML frontmatter. Images in the post have been added as attachments in the frontmatter and are also automatically downloaded. Image urls and internal links are re-written to be relative (but rooted).